+++ test/config/env-hostname.js
@@ -1,0 +1,3 @@
+'use strict'
+
+module.exports = process.env.TEST_HOSTNAME || 'env-hostname-default'