+++ test/config/js-fallback.ini.js
@@ -1,0 +1,7 @@
+'use strict'
+
+module.exports = {
+  main: {
+    host: process.env.TEST_JS_HOST || 'js-ini-fallback-host',
+  },
+}